backend/requirements.txt
fastapi>=0.115.0,<1.0
uvicorn[standard]>=0.32.0,<1.0
python-dotenv>=1.0.0,<2.0
httpx>=0.27.0,<1.0
python-multipart>=0.0.9,<1.0
faster-whisper>=1.0.3,<2.0
pytest>=8.0,<9.0
